Move FunctionResourceLoader into SessionState
