+++ sql/hive/src/main/scala/org/apache/spark/sql/hive/HiveSessionState.scala
@@ -87,7 +87,7 @@ private[hive] class HiveSessionState(ctx: SQLContext) extends SessionState(ctx) 
       sharedState.externalCatalog,
       metadataHive,
       ctx,
-      ctx.functionResourceLoader,
+      functionResourceLoader,
       functionRegistry,
       conf,
       hiveconf)
